Make auto-update more reliable
the $(shell ...) expansion in Makefile didn't work like I thought, they
are expanded when reading the file not when executing. As such they never
worked correctly.

This adjusts the Makefile to work in a more standard fashion, comparing the
timestamps between the `setup.sh` file and a `provision` file that gets
touched on every run to decide if we need to re-provision.

Change-Id: I6c434bc178cb770ab40a4403fc1a68faf9ad9714